requirements_freeze.txt
anyio==3.6.1
certifi==2022.6.15
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.5
commonmark==0.9.1
distlib==0.3.6
editables==0.3
filelock==3.8.0
ftfy==6.1.1
h11==0.12.0
hatch==1.5.0
hatchling==1.8.1
httpcore==0.15.0
httpx==0.23.0
hyperlink==21.0.0
idna==3.3
jaraco.classes==3.2.2
keyring==23.9.1
mailchecker==5.0.0
more-itertools==8.14.0
packaging==21.3
pathspec==0.10.1
pexpect==4.8.0
phonenumbers==8.12.54
platformdirs==2.5.2
pluggy==1.0.0
ptyprocess==0.7.0
Pygments==2.13.0
pyparsing==3.0.9
pyperclip==1.8.2
python-benedict==0.25.4
python-dateutil==2.8.2
python-fsutil==0.6.1
python-slugify==6.1.2
pywin32-ctypes==0.2.0
PyYAML==6.0
requests==2.28.1
rfc3986==1.5.0
rich==12.5.1
shellingham==1.5.0
six==1.16.0
sniffio==1.3.0
text-unidecode==1.3
toml==0.10.2
tomli==2.0.1
tomli_w==1.0.0
tomlkit==0.11.4
urllib3==1.26.12
userpath==1.8.0
virtualenv==20.16.4
wcwidth==0.2.5
xmltodict==0.13.0
